+++ hilt/annotations/src/main/kotlin/Primary.kt
@@ -22,8 +22,9 @@ import kotlin.reflect.KClass
 /**
  * Marks primary implementation of the given [supertype].
  *
- * The implementation will be bound to the given supertype in the given [component]. Optionally, it will be annotated
- * with the given [scope] and/or [qualifier].
+ * The implementation will be bound to the given supertype in the given [component].
+ *
+ * Any annotations annotated with `@Scope` or `@Qualifier` will also annotate the resulting `@Binds` method.
  */
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.SOURCE)
@@ -42,19 +43,5 @@ annotation class Primary(
      *
      * Must be a type annotated with `@DefineComponent`.
      */
-    val component: KClass<*>,
-
-    /**
-     * Scope for which the primary implementation is provided.
-     *
-     * Must be an annotation class annotated with `@Scope`.
-     */
-    val scope: KClass<*> = Unscoped::class,
-
-    /**
-     * Scope for which the primary implementation is provided.
-     *
-     * Must be an annotation class annotated with `@Qualifier`.
-     */
-    val qualifier: KClass<*> = Unqualified::class
+    val component: KClass<*>
 )