+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Changed
 - **BREAKING CHANGE:** Rename `superclass` to `supertype`.
+- **BREAKING CHANGE:** Copy `@Scope` and `@Qualifier` annotations from the `@Primary`-annotated type
+  to the binding method instead of `scope` and `qualifier` annotation parameters.
 
 ## [1.0.0-ALPHA1]
 ### Added